Inclusão de Request no Módulo de Receitas

+++ app/Http/Controllers/ReceitaController.php
@@ -2,8 +2,7 @@
 
 namespace App\Http\Controllers;
 
-use App\Models\User;
-use Illuminate\Http\Request;
+use App\Http\Requests\ReceitaRequest;
 use Illuminate\View\View;
 use App\Models\Receita;
 use Illuminate\Http\RedirectResponse;
@@ -12,6 +11,8 @@ class ReceitaController extends Controller
 {
     /**
      * Listagem da receita do usuário
+     *
+     * @return View
      */
     public function index() : View
     {
@@ -24,6 +25,8 @@ class ReceitaController extends Controller
 
     /**
      * Formulário de inserção de receitas
+     *
+     * @return View
      */
     public function create() : View
     {
@@ -32,47 +35,58 @@ class ReceitaController extends Controller
 
     /**
      * Processa os dados do formulário de inserção de receita
+     *
+     * @param ReceitaRequest $request Injeção de Dependência do Laravel
+     * @return RedirectResponse
      */
-    public function store(Request $request) : RedirectResponse
+    public function store(ReceitaRequest $request) : RedirectResponse
     {
-        $validated = $request->validate([
-            'empresa' => 'required',
-            'valor' => 'required'
-        ]);
+        $validated = $request->safe()->all();
 
-        $validated['valor'] = str_replace('.', '', $validated['valor']);
-        $validated['valor'] = str_replace(',', '.', $validated['valor']);
+        $validated['valor'] = str_replace(',', '.', str_replace('.', '', $validated['valor']));
         $validated['user_id'] = auth()->user()->id;
 
-        $receita = new Receita($validated);
-
-        $receita->save();
+        Receita::create($validated);
 
         return redirect(route('receitas.index'))->with('msg', ['type' => 'success', 'msg' => 'Receita inserida com sucesso!']);
     }
 
     /**
      * Carrega o formulário para edição da receita
+     *
+     * @param int $id_receita Código da Receita
+     * @return View|RedirectResponse
      */
-    public function edit($id): View
+    public function edit(int $id_receita): View|RedirectResponse
     {
-        $receita = Receita::findOrFail($id);
+        $receita = Receita::where('id_receita', '=', $id_receita)->where('user_id', '=', auth()->user()->id)->first();
+
+        if(empty($receita))
+            return redirect(route('receitas.index'))->with('msg', ['type' => 'danger', 'msg' => 'Receita informada não encontrada']);
+
         $receita['valor'] = number_format($receita['valor'], 2, ',', '.');
 
-        return view('receitas/form', ['receita' => $receita, 'activeReceitas' => 'active']);
+        return view('receitas.form', ['receita' => $receita, 'activeReceitas' => 'active']);
     }
 
     /**
      * Recebe os dados do formulário de edição de receitas e salva
+     *
+     * @param int $id_receita Código da Receita
+     * @param ReceitaRequest $request Injeção de Dependência do Laravel
+     * @return RedirectResponse
      */
-    public function update($id, Request $request): RedirectResponse
+    public function update(int $id_receita, ReceitaRequest $request): RedirectResponse
     {
-        $receita = Receita::findOrFail($id);
+        $receita = Receita::where('id_receita', '=', $id_receita)->where('user_id', '=', auth()->user()->id)->first();
 
-        $receita->empresa = $request->input('empresa');
-        $valor = str_replace('.', '', $request->valor);
-        $valor = str_replace(',', '.', $valor);
-        $receita->valor = $valor;
+        if(empty($receita))
+            return redirect(route('receitas.index'))->with('msg', ['type' => 'danger', 'msg' => 'Receita informada não encontrada']);
+
+        $validated = $request->safe()->all();
+
+        $receita->empresa = $validated['empresa'];
+        $receita->valor = str_replace(',', '.', str_replace('.', '', $validated['valor']));
 
         $receita->update();
 
@@ -81,12 +95,19 @@ class ReceitaController extends Controller
 
     /**
      * Remove a receita informada
+     *
+     * @param int $id_receita Código da Receita
+     * @return RedirectResponse
      */
-    public function destroy($id): RedirectResponse
+    public function destroy(int $id_receita): RedirectResponse
     {
-        $receita = Receita::findOrFail($id);
+        $receita = Receita::where('id_receita', '=', $id_receita)->where('user_id', '=', auth()->user()->id)->first();
+
+        if(empty($receita))
+            return redirect(route('receitas.index'))->with('msg', ['type' => 'danger', 'msg' => 'Receita informada não encontrada']);
 
         $receita->delete();
+
         return redirect(route('receitas.index'))->with('msg', ['type' => 'success', 'msg' => 'Receita removida.']);
     }
 }
